Rename private route to protect route
